requirements.txt
numpy
ultralytics==8.3.0
opencv-python-headless==4.10.0.84
Pillow
streamlit